+++ ui/components/head-to-head/player-row.tsx
@@ -1,5 +1,4 @@
 import * as React from 'react';
-import { Grid, Row, Col } from 'react-bootstrap';
 import { Player } from 'tntfl-api';
 
 import PlayerName from '../player-name';
@@ -15,33 +14,28 @@ interface PlayerRowProps {
 }
 export default function PlayerRow(props: PlayerRowProps): JSX.Element {
   const { player1Name, player2Name, player1, player2, numActivePlayers, base } = props;
-  const columnStyle = {padding: 0};
   return (
     <tr>
       <td style={{width: '30%', padding: 0}}>
-        <Grid fluid={true}>
-          <Row>
-            <Col xs={8} style={columnStyle}>
-              <PlayerName name={player1Name} base={base}/>
-            </Col>
-            <Col xs={4} style={columnStyle}>
-              {player1 ? <Rank rank={player1.rank} numActivePlayers={numActivePlayers}/> : 'Loading...'}
-            </Col>
-          </Row>
-        </Grid>
+        <div style={{display: 'flex'}}>
+          <div style={{width: '70%'}}>
+            <PlayerName name={player1Name} base={base} colour={'red-player'}/>
+          </div>
+          <div style={{width: '30%'}}>
+            {player1 ? <Rank rank={player1.rank} numActivePlayers={numActivePlayers}/> : 'Loading...'}
+          </div>
+        </div>
       </td>
       <td/>
       <td style={{width: '30%', padding: 0}}>
-        <Grid fluid={true}>
-          <Row>
-            <Col xs={4} style={columnStyle}>
-              {player2 ? <Rank rank={player2.rank} numActivePlayers={numActivePlayers}/> : 'Loading...'}
-            </Col>
-            <Col xs={8} style={columnStyle}>
-              <PlayerName name={player2Name} base={base}/>
-            </Col>
-          </Row>
-        </Grid>
+        <div style={{display: 'flex'}}>
+          <div style={{width: '30%'}}>
+            {player2 ? <Rank rank={player2.rank} numActivePlayers={numActivePlayers}/> : 'Loading...'}
+          </div>
+          <div style={{width: '70%'}}>
+            <PlayerName name={player2Name} base={base} colour={'blue-player'}/>
+          </div>
+        </div>
       </td>
     </tr>
   );