+++ ui/components/ladder-panel.tsx
@@ -7,7 +7,8 @@ import Ladder from './ladder';
 
 interface LadderPanelProps extends Props<LadderPanel> {
   entries?: LadderEntry[];
-  bsStyle?: string;
+  speculative?: boolean;
+  base?: string;
 }
 interface State {
   showInactive: boolean;
@@ -22,13 +23,14 @@ export default class LadderPanel extends Component<LadderPanelProps, State> {
     this.setState({showInactive: !showInactive});
   }
   render(): JSX.Element {
-    const { entries, bsStyle } = this.props;
+    const { entries, speculative, base = '' } = this.props;
     const { showInactive } = this.state;
+    const style = speculative ? { backgroundColor: '#faebcc' } : undefined;
     return (
-      <Panel bsStyle={bsStyle}>
+      <Panel style={style}>
         {entries
           ? <div>
-              <Ladder entries={entries} showInactive={showInactive}/>
+              <Ladder entries={entries} showInactive={showInactive} style={style} base={base}/>
               <Button onClick={() => this.onShowInactive()} style={{width: '100%'}}>
                 {showInactive ? 'Hide inactive' : 'Show inactive'}
               </Button>