+++ ui/components/stats/achievements.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Col, Panel } from 'react-bootstrap';
+import { Panel } from 'react-bootstrap';
 import { Achievement, AchievementCount } from 'tntfl-api';
 
 interface AchievementPanelProps {
@@ -10,9 +10,12 @@ function AchievementPanel(props: AchievementPanelProps): JSX.Element {
   const { achievement, count } = props;
   const icon = 'achievement-' + achievement.name.replace(/ /g, '');
   return (
-    <Panel header={achievement.name} style={{textAlign: 'center'}}>
-      <div className={icon} style={{margin: 'auto'}}/>
-      {achievement.description} - <b>{count}</b>
+    <Panel>
+      <Panel.Heading style={{textAlign: 'center'}}>{achievement.name}</Panel.Heading>
+      <Panel.Body>
+        <div className={icon} style={{margin: 'auto'}}/>
+        {achievement.description} - <b>{count}</b>
+      </Panel.Body>
     </Panel>
   );
 }
@@ -23,8 +26,13 @@ interface AchievementsProps {
 export default function Achievements(props: AchievementsProps): JSX.Element {
   const { achievements } = props;
   return (
-    <Panel header={'Achievements'}>
-      {achievements.map((a, i) => <Col xs={3} key={`${i}`}><AchievementPanel achievement={a} count={a.count}/></Col>)}
+    <Panel>
+      <Panel.Heading>Achievements</Panel.Heading>
+      <Panel.Body>
+        <div style={{display: 'grid', gridTemplateColumns: 'repeat(4, auto)', gridColumnGap: 20}}>
+          {achievements.map((a, i) => <div key={`${i}`}><AchievementPanel achievement={a} count={a.count}/></div>)}
+        </div>
+      </Panel.Body>
     </Panel>
   );
 }