+++ ui/components/stats/records-section.tsx
@@ -14,17 +14,20 @@ export default function RecordsSection(props: RecordsSectionProps): JSX.Element 
   const { winningStreak, longestGame } = records;
   const longestGameGoals = longestGame.red.score + longestGame.blue.score;
   return (
-    <Panel header={'Records'}>
-      <dl className='dl-horizontal'>
-        <StatListItem
-          name={'Longest winning streak'}
-          value={<span><b>{winningStreak.count}</b> (<PlayerLink name={winningStreak.player} base={base}/>)</span>}
-        />
-        <StatListItem
-          name={'Longest game'}
-          value={<span><b><a href={`${base}game/${longestGame.date}`}>{longestGameGoals} goals</a></b> (<PlayerLink name={longestGame.red.name} base={base}/> vs <PlayerLink name={longestGame.blue.name} base={base}/>)</span>}
-        />
-      </dl>
+    <Panel>
+      <Panel.Heading>Records</Panel.Heading>
+      <Panel.Body>
+        <dl className='dl-horizontal'>
+          <StatListItem
+            name={'Longest winning streak'}
+            value={<span><b>{winningStreak.count}</b> (<PlayerLink name={winningStreak.player} base={base}/>)</span>}
+          />
+          <StatListItem
+            name={'Longest game'}
+            value={<span><b><a href={`${base}game/${longestGame.date}`}>{longestGameGoals} goals</a></b> (<PlayerLink name={longestGame.red.name} base={base}/> vs <PlayerLink name={longestGame.blue.name} base={base}/>)</span>}
+          />
+        </dl>
+      </Panel.Body>
     </Panel>
   );
 }