+++ ui/containers/game.tsx
@@ -1,15 +1,15 @@
 import * as React from 'react';
 import { Component, Props } from 'react';
-import { Grid, Row, Panel } from 'react-bootstrap';
 import * as ReactDOM from 'react-dom';
-import { GamesApi, Game } from 'tntfl-api';
-import 'whatwg-fetch';
+import { Panel } from 'react-bootstrap';
+import { Game, GamesApi, PlayersApi } from 'tntfl-api';
+import '../styles/achievement.less';
+import '../styles/style.less';
 
 import GameSummary from '../components/game-summary';
 import GameDetails from '../components/game-details';
 import NavigationBar from '../components/navigation-bar';
 import { getParameters } from '../utils/utils';
-import { PlayersApi } from '../../swagger/tntfl-api/dist/api';
 
 interface GamePageProps extends Props<GamePage> {
   base: string;
@@ -59,16 +59,18 @@ class GamePage extends Component<GamePageProps, GamePageState> {
           base={base}
         />
         {game ?
-          <Grid fluid={true}>
+          <div className={'page-container'}>
             <Panel>
-              <Row>
-                <GameSummary game={game} base={'../../'} numActivePlayers={numActivePlayers} />
-              </Row>
-              <Row>
-                <GameDetails game={game} punditry={punditry}/>
-              </Row>
+              <Panel.Body>
+                <div>
+                  <GameSummary game={game} base={'../../'} numActivePlayers={numActivePlayers} />
+                </div>
+                <div>
+                  <GameDetails game={game} punditry={punditry}/>
+                </div>
+              </Panel.Body>
             </Panel>
-          </Grid>
+          </div>
           : 'Loading...'
         }
       </div>
@@ -78,7 +80,7 @@ class GamePage extends Component<GamePageProps, GamePageState> {
 
 ReactDOM.render(
     <GamePage
-      base={'../../'}
+      base={__tntfl_base_path__}
       gameId={getParameters(1)[0]}
     />,
     document.getElementById('entry'),