+++ ui/containers/player.tsx
@@ -1,8 +1,11 @@
 import * as React from 'react';
 import { Component, Props } from 'react';
-import { Panel, Grid, Row, Col } from 'react-bootstrap';
+import { Panel } from 'react-bootstrap';
 import * as ReactDOM from 'react-dom';
-import { Game, Achievement, Player } from 'tntfl-api';
+import { Achievement, Game, Player, PlayersApi } from 'tntfl-api';
+import 'react-bootstrap-table/css/react-bootstrap-table.css';
+import '../styles/achievement.less';
+import '../styles/style.less';
 
 import PerPlayerStats from './per-player-stats';
 import PlayerAchievements from '../components/player/player-achievements';
@@ -11,7 +14,6 @@ import PlayerStats from '../components/player/player-stats';
 import RecentGames from '../components/recent-game-list';
 import NavigationBar from '../components/navigation-bar';
 import { getParameters, mostRecentGames } from '../utils/utils';
-import { PlayersApi } from '../../swagger/tntfl-api/dist/api';
 
 interface PlayerPageProps extends Props<PlayerPage> {
   base: string;
@@ -68,26 +70,30 @@ class PlayerPage extends Component<PlayerPageProps, PlayerPageState> {
           base={base}
         />
         {player && games ?
-          <Grid fluid={true}>
-            <Row>
-              <Col md={8}>
-                <PlayerStats player={player} numActivePlayers={activePlayers || 0} games={games} base={base}/>
-                <Panel header={<h2>Skill Chart</h2>}>
+          <div className={'ladder-page'}>
+            <div className={'ladder-panel'}>
+              <PlayerStats player={player} numActivePlayers={activePlayers || 0} games={games} base={base}/>
+              <Panel>
+                <Panel.Heading><h2>Skill Chart</h2></Panel.Heading>
+                <Panel.Body>
                   <PlayerSkillChart playerName={player.name} games={games} />
-                </Panel>
-                <PerPlayerStats playerName={playerName} base={base}/>
-              </Col>
-              <Col md={4}>
-                <RecentGames games={mostRecentGames(games)} showAllGames={true} base={base}/>
-                <Panel header={<h2>Achievements</h2>}>
+                </Panel.Body>
+              </Panel>
+              <PerPlayerStats playerName={playerName} base={base}/>
+            </div>
+            <div className={'side-panel'}>
+              <RecentGames games={mostRecentGames(games)} showAllGames={true} base={base}/>
+              <Panel>
+                <Panel.Heading><h2>Achievements</h2></Panel.Heading>
+                <Panel.Body>
                   {achievements
                     ? <PlayerAchievements achievements={achievements} base={base}/>
                     : 'Loading...'
                   }
-                </Panel>
-              </Col>
-            </Row>
-          </Grid>
+                </Panel.Body>
+              </Panel>
+            </div>
+          </div>
           : 'Loading...'
         }
       </div>
@@ -97,7 +103,7 @@ class PlayerPage extends Component<PlayerPageProps, PlayerPageState> {
 
 ReactDOM.render(
   <PlayerPage
-    base={'../../'}
+    base={__tntfl_base_path__}
     playerName={getParameters(1)[0]}
   />,
   document.getElementById('entry'),