+++ ui/path/path1.ts
@@ -1,0 +1,2 @@
+__tntfl_base_path__ = '../';
+__webpack_public_path__ = `${__tntfl_base_path__}dist/`;